Expanded error handling to accommodate for wifi outages or lost connections
Previously, whenever my wifi connection was unstable the board would error reliably within an hour. Hopefully this will solve this by catching any and all errors and resetting eventually if errors cannot be resolved naturally. It is not the most precise or efficient way of handling errors, but optimizes for ease and fewer headaches for a dashboard on which my life does not depend.